Remove no-tls1 option
Looks like it was added back in 2013 to support some "broken servers":
https://sourceforge.net/p/fdm/mailman/message/30814502/
TLSv1.0 is insecure anyway and should always be disabled.
This change is planned for subsequent patches.
Also, this option does not disable TLSv1.1 and TLSv1.2.

For reference, the commit that adds no-tls1 option is
d0344971b5fdde41c52dfa57e472a6661b56ec30