XDG base directory: lock file in XDG_RUNTIME_DIR
Continuation of https://github.com/nicm/fdm/issues/96 implementation.

From https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html:

> There is a single base directory relative to which user-specific
> runtime files and other file objects should be placed.  This directory
> is defined by the environment variable $XDG_RUNTIME_DIR.

This also may provide slight performance benefits to this as
XDG_RUNTIME_DIR is often set to a directory under /tmp which some
systems configure to be mounted as tmpfs.

If XDG_RUNTIME_DIR is not set or is an empty string, the old behaviour
of the user's home directory is used.